Updated Import Logic To Support Django 4
In order to enable compatibility with Django 4, two specific imports
needed to be updated: gettext_lazy and url, both of which were deprecated
in Django 4.

To allow for backwards compatibility, a try/except statement has been
included which first attempts to import the Django 4 alternatives
(ugettext_lazy and re_path) and then falls back to the Django 2/3
forms if those imports are not availablity.

Additionally, I've included some logic within the most recent migration
to alleviate the errors seen in Issue 29.